backend/pyproject.toml
[project]
name = "anki-image-cloze-backend"
version = "0.1.0"
description = "Local prototype backend for image occlusion Anki workflows."
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.116.0",
  "httpx>=0.28.0",
  "pillow>=11.0.0",
  "pydantic-settings>=2.10.0",
  "python-multipart>=0.0.20",
  "uvicorn[standard]>=0.35.0",
]

[project.optional-dependencies]
ocr = ["paddleocr>=2.8.0"]
vision = ["opencv-python-headless>=4.12.0.88"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
